Use the warn-once logger provider as the default logger provider
Previously, the noop provider was used as the default provider,
which meant that if a logger provider was not set, loggers would not
output anything. The noop provider implementation remains so that
users can explicitly set the default logger provider to the noop one
if that is the behavior they want.